Cite a package

Default output is rmarkdown format.

cite_package(
  package,
  brack = TRUE,
  startText = "",
  endText = "",
  bib_df = "refs",
  bib_file = "packages.bib",
  sep = ";"
)

Arguments

package
Character. Name of package.

brack
Logical. If true, include brackets - '('ref')' - in output.

startText
Character. Text to include before reference.

endText
Character. Text to include after reference.

bib_df
Dataframe of package citations, usually read in via bib2df::bib2df()

bib_file
Path to bibliography. Will be created if it doesn't exist. knitr::write_bib().

sep
Character. What character to use to separate more than one reference. Defaults to ';' which is needed for rmarkdown.

Value

Character string with appropriate markdown formatting to insert the appropriate reference at that point in the text. If bib_df did not exist when the function is called, bib_df is created in the global environment and if bib_file did not exist it is created.

Details

bib_df and bib_file will be created if they don't currently exist.

Examples

if (FALSE) {
library(envReport)
bib_file = tempfile()
cite_package("dplyr", bib_file = bib_file)
refs
rm(refs)
unlink(bib_file)
}