Map points from a data frame with coordinates in columns x and y

Map points from a data frame with coordinates in columns x and y

data_name_map(
  df,
  x = "long",
  y = "lat",
  polys = NULL,
  polys_col = NULL,
  polys_name = NULL,
  polys_palette = NULL,
  polys_alpha = 1,
  aoi = NULL,
  in_crs = 4283,
  out_crs = 4326,
  dots_args = list(),
  layout_args = list()
)

Arguments

df

Dataframe with coordinate columns.

x

Character name of column with X coordinates (e.g. "long").

y

Character name of column with Y coordinates (e.g. "lat").

polys

Simple feature with polygons to map points on.

polys_col

Column in polys to include in map legend.

polys_name

Character name to give to polys legend title.

polys_palette

Vector of column colours to use for polys.

polys_alpha

Passed to tmap::tm_polygons() alpha argument.

aoi

Object on which sf::st_bbox will return a bounding box for the map output.

in_crs

Coordinate reference system of x and y in df. Can be given as any coordinate reference system recognised by sf::st_crs (e.g. EPSG code).

out_crs

Coordinate reference system of resulting map.

dots_args

Named list of arguments to pass to tmap::tm_dots().

layout_args

Named list of arguments to pass to tmap::tm_layout().

Value

tmap object