Make package bibliography, including tweaks for known package issues.

Makes use of both knitr::write_bib() (Xie 2021, 2015) and bib2df::bib2df() (Ottolinger 2019) .

make_bib(
  bib_file = "packages.bib",
  make_key = FALSE,
  is_package_bib = TRUE,
  pacs = NULL
)

Arguments

bib_file
Path to bibliography. Will be created if it doesn't exist.

make_key
Logical. Generate a new key using author and year.

is_package_bib
Logical. Is this a bibliography of package citations? Defaults to TRUE.

pacs
Character. Name of any R packages to add to bib_file.

Value

Tweaked bib_file and data frame of references (from bib2df::bib2df).

References

Ottolinger P (2019). bib2df: Parse a BibTeX File to a Data Frame. R package version 1.1.1, https://github.com/ropensci/bib2df.

Xie Y (2014). “knitr: A Comprehensive Tool for Reproducible Research in R.” In Stodden V, Leisch F, Peng RD (eds.), Implementing Reproducible Computational Research. Chapman and Hall/CRC. ISBN 978-1466561595, http://www.crcpress.com/product/isbn/9781466561595.

Xie Y (2015). Dynamic Documents with R and knitr, 2nd edition. Chapman and Hall/CRC, Boca Raton, Florida. ISBN 978-1498716963, https://yihui.org/knitr/.

Xie Y (2021). knitr: A General-Purpose Package for Dynamic Report Generation in R. R package version 1.33, https://yihui.org/knitr/.

Examples

if (FALSE) {
library(envReport)
bib_file <- tempfile()
make_bib(pacs = .packages())
refs
rm(refs)
unlink(bib_file)
}