Map points from a data frame with coordinates in columns x and y

Map points from a data frame with coordinates in columns x and y

data_name_map(
  df,
  x = "long",
  y = "lat",
  map_class = NULL,
  polys = NULL,
  polys_col = NULL,
  polys_name = NULL,
  polys_palette = NULL,
  aoi = NULL,
  in_crs = 4283,
  out_crs = 4326,
  ...
)

Arguments

df
Dataframe with coordinate columns.

x
Character name of column with X coordinates (e.g. "long").

y
Character name of column with Y coordinates (e.g. "lat").

map_class
Optional character name of column to include in map legend.

polys
Simple feature with polygons to map points on.

polys_col
Column in polys to include in map legend.

polys_name
Character name to give to polys legend title.

polys_palette
Vector of column colours to use for polys.

aoi
Object on which sf::st_bbox will return a bounding box for the map output.

in_crs
Coordinate reference system of x and y in df. Can be given as any coordinate reference system recognised by sf::st_crs (e.g. EPSG code).

out_crs
Coordinate reference system of resulting map.

...
tmap::tm_dots arguments. Note col, title, palette and legend.format are already defined within data_name_map

Value

tmap object